Partially revert "coppied over thomas' work"
This partially reverts commit 3b26157a8726686ee44ed08c7c96a040a3df9c88.

Reverting the portion of this commit that included some of the contents of a separate work in progress branch which was not ready to be merged yet.
It needs to be reverted because it creates a merge conflict with the branch it copied from now that the copied branch is actually ready to be merged.